backend/requirements_phase1.txt
httpx==0.27.0
feedparser==6.0.11
langdetect==1.0.9
praw==7.7.1
telethon==1.36.0
python-dotenv==1.0.1
transformers>=4.40.0
torch==2.6.0
spacy>=3.8.0
scipy>=1.14.0
numpy>=2.1.0
sentencepiece==0.2.1
sacremoses==0.1.1